Moved to integration test
