+++ testing/mysql/integration_test.go
@@ -15,12 +15,14 @@
 package mysql_test
 
 import (
+	"bytes"
 	"context"
 	"flag"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"testing"
 	"time"
@@ -92,10 +94,7 @@ func dropDatabase(t *testing.T, dbPath string) {
 	}
 }
 
-func prepareIntegrationTest(t *testing.T) string {
-	if databaseAdmin == nil {
-		t.Skip("Integration tests skipped")
-	}
+func prepareTmpDir(t *testing.T) string {
 	tmpdir, err := ioutil.TempDir(".", "int-test-")
 	if err != nil {
 		log.Fatal(err)
@@ -104,9 +103,10 @@ func prepareIntegrationTest(t *testing.T) string {
 }
 
 func TestIntegration_MYSQLDUMP_SimpleUse(t *testing.T) {
+	onlyRunForEmulatorTest(t)
 	t.Parallel()
 
-	tmpdir := prepareIntegrationTest(t)
+	tmpdir := prepareTmpDir(t)
 	defer os.RemoveAll(tmpdir)
 
 	now := time.Now()
@@ -132,7 +132,7 @@ func TestIntegration_MYSQL_SimpleUse(t *testing.T) {
 	onlyRunForEmulatorTest(t)
 	t.Parallel()
 
-	tmpdir := prepareIntegrationTest(t)
+	tmpdir := prepareTmpDir(t)
 	defer os.RemoveAll(tmpdir)
 
 	now := time.Now()
@@ -148,6 +148,62 @@ func TestIntegration_MYSQL_SimpleUse(t *testing.T) {
 	defer dropDatabase(t, dbPath)
 
 	checkResults(t, dbPath)
+}
+
+func schemaOnlyTest(t *testing.T, dbName, filePrefix, sessionFile, dumpFilePath string) {
+	cmd := exec.Command("bash", "-c", fmt.Sprintf("go run github.com/cloudspannerecosystem/harbourbridge -driver mysqldump -schema-only -dbname %s -prefix %s < %s", dbName, filePrefix, dumpFilePath))
+	var out, stderr bytes.Buffer
+	cmd.Stdout = &out
+	cmd.Stderr = &stderr
+	if err := cmd.Run(); err != nil {
+		fmt.Printf("stdout: %q\n", out.String())
+		fmt.Printf("stderr: %q\n", stderr.String())
+		t.Fatal(err)
+	}
+	if _, err := os.Stat(fmt.Sprintf("%sreport.txt", filePrefix)); os.IsNotExist(err) {
+		t.Fatalf("report file not generated during schema-only test")
+	}
+	if _, err := os.Stat(fmt.Sprintf("%sschema.ddl.txt", filePrefix)); os.IsNotExist(err) {
+		t.Fatalf("legal ddl file not generated during schema-only test")
+	}
+	if _, err := os.Stat(fmt.Sprintf("%sschema.txt", filePrefix)); os.IsNotExist(err) {
+		t.Fatalf("readable schema file not generated during schema-only test")
+	}
+	if _, err := os.Stat(sessionFile); os.IsNotExist(err) {
+		t.Fatalf("session file not generated during schema-only test")
+	}
+}
+
+func dataOnlyTest(t *testing.T, dbName, dbURI, filePrefix, sessionFile, dumpFilePath string) {
+	cmd := exec.Command("bash", "-c", fmt.Sprintf("go run github.com/cloudspannerecosystem/harbourbridge -driver mysqldump -data-only -instance %s -dbname %s -prefix %s -session %s < %s", instanceID, dbName, filePrefix, sessionFile, dumpFilePath))
+	var out, stderr bytes.Buffer
+	cmd.Stdout = &out
+	cmd.Stderr = &stderr
+	cmd.Env = append(os.Environ(),
+		fmt.Sprintf("GCLOUD_PROJECT=%s", projectID),
+	)
+	if err := cmd.Run(); err != nil {
+		fmt.Printf("stdout: %q\n", out.String())
+		fmt.Printf("stderr: %q\n", stderr.String())
+		t.Fatal(err)
+	}
+	defer dropDatabase(t, dbURI)
+	checkResults(t, dbURI)
+}
+
+func TestMainApp(t *testing.T) {
+	tmpdir := prepareTmpDir(t)
+	defer os.RemoveAll(tmpdir)
+
+	dbName := "test-individual-flows"
+	dumpFilePath := "../../test_data/mysqldump.test.out"
+	filePrefix := filepath.Join(tmpdir, dbName+".")
+	sessionFile := fmt.Sprintf("%ssession.json", filePrefix)
+	schemaOnlyTest(t, dbName, filePrefix, sessionFile, dumpFilePath)
+	// Skip data only test if emulator is not running.
+	onlyRunForEmulatorTest(t)
+	dbURI := fmt.Sprintf("projects/%s/instances/%s/databases/%s", projectID, instanceID, dbName)
+	dataOnlyTest(t, dbName, dbURI, filePrefix, sessionFile, dumpFilePath)
 }
 
 func checkResults(t *testing.T, dbPath string) {
